requirements.txt
opencv-python
yt-dlp
ffmpeg-python
PyAudio